struct(T8): extract deterministic RNG streams into rng.rs
Document and house the world/action stream builders in one module: shared
splitmix64 mixer, world_rng_for_week, action_rng_for_week, and Game::action_rng*.
Salts remain in constants.rs. turn.rs uses the world builder; core keeps
lifecycle only. Pure move — draw order unchanged; determinism tests pass.
